feat: add DecodeResult class to ProtoDecoder
Add a new class `DecodeResult` to `ProtoDecoder` in the `protobuf.magic` package. The `DecodeResult` class contains a list of `Part` objects and a byte array `leftOver`. This class is used to represent the result of decoding a protobuf message.

This commit also includes changes to the `VarintUtils` class where the `decodeVarint` method now returns a `VarintResult` instead of a `VarIntResult`.

Note: This commit includes multiple changes in different files.