feat: Add Protobuf class
This commit adds a new `Protobuf` class that represents a protobuf message. The `Protobuf` class has two fields: `type` which represents the type of the protobuf message, and `value` which represents the value of the message. The `Protobuf` class also provides getter methods for accessing the type and value.

The purpose of this class is to provide a more structured representation of protobuf messages and allow